packages/artifact-core-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "primust-artifact-core"
version = "0.0.1"
description = "Canonical JSON, hashing, signing (Python mirror)"
requires-python = ">=3.11"
license = "LicenseRef-Proprietary"
dependencies = [
  "pynacl>=1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/primust_artifact_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
